double instead of long in exact eigvals. Added turned blocks

+++ exact-diag-wrapper.py
@@ -11,7 +11,7 @@ def exactdiags_cpp(N=6, samples=100, Nb_values=50):
     CN=ctypes.c_int(N)
     Csamples=ctypes.c_int(samples)
     CNb_values = ctypes.c_int(Nb_values)
-    values = (ctypes.c_float*(samples*Nb_values))()
+    values = (ctypes.c_double*(samples*Nb_values))()
     diag.computePackSpectrum(N, samples, Nb_values, values)
     print(time.time()-t)
     return values
@@ -21,7 +21,7 @@ def exactdiag_cpp(J=0.1, N=6, Nb_values=50):
     CN=ctypes.c_int(N)
     CJ=ctypes.c_float(J)
     CNb_values = ctypes.c_int(Nb_values)
-    values = (ctypes.c_float*Nb_values)()
+    values = (ctypes.c_double*Nb_values)()
     diag.computeSpectrum(CN,CJ,CNb_values, values)
     print(time.time()-t)
     return values
